Fix Optimizer based on new passenger generating

+++ src/main/java/put/plane/boarding/optimizing/OptimizerResult.java
@@ -3,6 +3,6 @@ package put.plane.boarding.optimizing;
 import java.util.List;
 
 public record OptimizerResult(
-        int bestTime,
+        float bestTime,
         List<Integer> bestSolution
 ) {}
